# docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Simple User Post",
      "description": "Aplikasi Untuk Posting",
      "termsOfService": "http://alif.com/terms/",
      "contact": {
        "name": "API Support",
        "url": "http://www.alif.com/support",
        "email": "support@example.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0.1"
    },
    "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Development Server"
        },
        {
          "url": "http://production.com",
          "description": "Production Server"
        }
      ],
      "tags": [
        {
          "name": "Player",
          "description": "End Point Untuk Operasi CRUD Player"
        }
      ],
      "paths" : {
        "/api/v1/players": {
            "post": {
              "tags": ["Player"],
              "summary": "Create A new Player",
              "operationId": "createPlayer",
              "requestBody": {
                "description": "Buat Player Baru",
                "content": {
                  "application/x-www-form-urlencoded": {
                    "schema": {
                      "$ref": "#/components/schemas/Create_Player_Body"
                    }
                  },
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Create_Player_Body"
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "Success",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {
                            "$ref": "#/components/schemas/Player"
                          }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "Failed",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example" : "username or email cannot empty || password cannot be empty"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Failed",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Internal server error"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "get": {
              "tags": ["Player"],
              "summary": "Get All Player",
              "operationId": "getAllPlayer",
              "description": "Get All Playaer From Database",
              "responses": {
                "200": {
                  "description": "Successfully get Player",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Success"
                          },
                          "data": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/Player"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Failed",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Internal server error"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/v1/players/{id}": {
            "get": {
              "tags": ["Player"],
              "summary": "Get Player By ID",
              "operationId": "getPlayerByID",
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "description": "Unique ID of User",
                  "required": true,
                  "schema": {
                    "type": "integer",
                    "example": "1"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Successfully get Player",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Success"
                          },
                          "data": {
                            "$ref": "#/components/schemas/Player"
                          }
                        }
                      }
                    }
                  }
                },
                "404": {
                  "description": "Player Data Not Found",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Player with ID not found"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Failed",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Internal server error"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "put": {
              "tags": ["Player"],
              "summary": "Edit Player By ID",
              "operationId": "editPlayerByID",
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "description": "Unique ID of Player",
                  "required": true,
                  "schema": {
                    "type": "integer",
                    "example": "1"
                  }
                }
              ],
              "requestBody": {
                "description": "Update Existing Player Data",
                "content": {
                  "application/x-www-form-urlencoded": {
                    "schema": {
                      "$ref": "#/components/schemas/Update_Player_Body"
                    }
                  },
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Update_Player_Body"
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "Success",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Player with id successfully updated"
                          },
                          "data": {
                            "$ref": "#/components/schemas/Player"
                          }
                        }
                      }
                    }
                  }
                },
                "404": {
                  "description": "Player Data Not Found",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Player Data Not Found"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Failed",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Failed to Update"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "delete": {
              "tags": ["Player"],
              "summary": "Delete Player By ID",
              "operationId": "deletePlayerByID",
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "description": "Unique ID of User",
                  "required": true,
                  "schema": {
                    "type": "string",
                    "example": "1"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Success",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Player with id, was deleted successfully"
                          }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "FAILED",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Cannot delete Player with id. Maybe Player was not found!"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Failed",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Internal server error"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/v1/players/exp/{id}":{
            "post": {
                "tags": ["Player"],
                "summary": "Edit Exp By ID",
                "operationId": "editExpByID",
                "parameters": [
                  {
                    "name": "id",
                    "in": "path",
                    "description": "Unique ID of Player",
                    "required": true,
                    "schema": {
                      "type": "integer",
                      "example": "1"
                    }
                  }
                ],
                "requestBody": {
                  "description": "Update Existing Exp Player Data",
                  "content": {
                    "application/x-www-form-urlencoded": {
                      "schema": {
                        "$ref": "#/components/schemas/Update_Exp_Player_Body"
                      }
                    },
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/Update_Exp_Player_Body"
                      }
                    }
                  }
                },
                "responses": {
                  "200": {
                    "description": "SUCCESS",
                    "content": {
                      "application/json": {
                        "schema": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "example": "Player with id = ${id} gain ${exp} experience. Total experience = ${player.experience}"
                            }
                          }
                        }
                      }
                    }
                  },
                  "404": {
                    "description": "Not found",
                    "content": {
                      "application/json": {
                        "schema": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "example": "Player with ${id} not found"
                            }
                          }
                        }
                      }
                    }
                  },
                  "400": {
                    "description": "Failed",
                    "content": {
                      "application/json": {
                        "schema": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "example": "Cannot update Player with id=${id}!"
                            }
                          }
                        }
                      }
                    }
                  },
                  "500": {
                    "description": "Failed",
                    "content": {
                      "application/json": {
                        "schema": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "example": "Internal server error"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
          }
      },
      "components" : {
        "schemas": {
            "Create_Player_Body": {
                "required": ["username", "email", "password", "experience"],
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "Alif"
                  },
                  "email": {
                    "type": "string",
                    "example": "alif@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "alif"
                  },
                  "experience": {
                    "type": "integer",
                    "example": "100000"
                  }
                }
              },
              "Update_Player_Body": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "Alif"
                  },
                  "email": {
                    "type": "string",
                    "example": "alif@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "alif"
                  },
                  "experience": {
                    "type": "integer",
                    "example": "100000"
                  }
                }
              },
              "Update_Exp_Player_Body": {
                "type": "object",
                "properties": {
                  "exp": {
                    "type": "integer",
                    "example": "1000000"
                  }
                }
              },
              "Player": {
                "required": ["username", "email", "password", "experience", "lvl"],
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "experience": {
                    "type": "integer"
                  },
                  "lvl": {
                    "type": "integer"
                  },
                  "created_at": {
                    "type": "string",
                    "format": "date"
                  },
                  "updated_at": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }  
                
        }
    }
}